Fix debug launch failures from variable resolution
The debug session manager relies on the variable resolving services
to replace variables present in the launch configuration.

Some of these variables are resolved by interactive commands towards
the user, however if the user press escape or ignores a prompt, the
variable resolution will not take place and if the flow continues,
errors will be displayed to the user (unnecessarily).

The solution before this fix was relying on a check on the successful
resolution of all the variables i.e. none are 'undefined', however
there are cases were this happens and the flow shall continue.

This fix, replaces the check for the presence of `escape` entries in
an array, this array is populated with the command ids that were not
successfully resolved i.e. giving a result of 'null'.

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>